Added the conda-gitenv subcommands into a common cli. Moved cli to a better name (resolve).
