+++ setup.py
@@ -11,7 +11,8 @@ setup(
       packages=['conda_env_tracker'],
       entry_points={
           'console_scripts': [
-              'conda-env-tracker = conda_env_tracker.cli:main',
+              'conda-gitenv = conda_env_tracker.cli:main',
+              'conda-env-tracker = conda_env_tracker.resolve:main',
               'conda-env-tracker-timestamp = conda_env_tracker.tag_dates:main',
               'conda-env-tracker-labeltag = conda_env_tracker.label_tag:main',
               'conda-env-tracker-deploy = conda_env_tracker.deploy:main',
